requirements.txt
openai
ollama
faiss-cpu
google-cloud-storage
langchain